Refactored TabLineFactory into ProblemFactory.
As part of https://github.com/processing/processing4/pull/1, refactored TabLineFactory into ProblemFactory, merging tests in the process.
