+++ java/src/processing/mode/java/pdex/util/ProblemFactory.java
@@ -31,7 +31,7 @@ public class ProblemFactory {
     }
 
     // Get local area
-    TabLine tabLine = TabLineFactory.getTab(tabStarts, line);
+    TabLine tabLine = getTab(tabStarts, line);
 
     int tab = tabLine.getTab();
     int localLine = tabLine.getLineInTab(); // Problems emitted in 0 index
@@ -65,7 +65,7 @@ public class ProblemFactory {
   public static Problem build(PdePreprocessIssue pdePreprocessIssue, List<Integer> tabStarts) {
     int line = pdePreprocessIssue.getLine();
 
-    TabLine tabLine = TabLineFactory.getTab(tabStarts, line);
+    TabLine tabLine = getTab(tabStarts, line);
 
     int tab = tabLine.getTab();
     int localLine = tabLine.getLineInTab();
@@ -84,6 +84,25 @@ public class ProblemFactory {
         localLine,
         localLine + col
     );
+  }
+
+  /**
+   * Get the local tab and line number for a global line.
+   *
+   * @param tabStarts The lines on which each tab starts.
+   * @param line The global line to locate as a local line.
+   * @return The local tab number and local line number.
+   */
+  protected static TabLine getTab(List<Integer> tabStarts, int line) {
+    OptionalInt tabMaybe = IntStream.range(0, tabStarts.size())
+        .filter((index) -> line >= tabStarts.get(index))
+        .max();
+
+    int tab = tabMaybe.orElse(0);
+
+    int localLine = line - tabStarts.get(tab);
+
+    return new TabLine(tab, line, localLine);
   }
 
 }
