+++ java/test/processing/mode/java/pdex/util/ProblemFactoryTest.java
@@ -18,6 +18,7 @@ public class ProblemFactoryTest {
   private PdePreprocessIssue pdePreprocessIssue;
   private List<Integer> tabStarts;
   private Editor editor;
+  private List<Integer> starts;
 
   @Before
   public void setUp() {
@@ -29,6 +30,11 @@ public class ProblemFactoryTest {
     editor = Mockito.mock(Editor.class);
     Mockito.when(editor.getLineStartOffset(3)).thenReturn(10);
     Mockito.when(editor.getLineStopOffset(3)).thenReturn(12);
+
+    starts = new ArrayList<>();
+    starts.add(0);
+    starts.add(5);
+    starts.add(10);
   }
 
   @Test
@@ -47,6 +53,33 @@ public class ProblemFactoryTest {
 
     Assert.assertEquals(3, problem.getLineNumber());
     Assert.assertEquals("test", problem.getMessage());
+  }
+
+  @Test
+  public void getTabStart() {
+    Assert.assertEquals(0, ProblemFactory.getTab(starts, 0).getTab());
+  }
+
+  @Test
+  public void getTabMiddleFrontEdge() {
+    Assert.assertEquals(1, ProblemFactory.getTab(starts, 5).getTab());
+  }
+
+  @Test
+  public void getTabMiddle() {
+    TabLine tabLine = ProblemFactory.getTab(starts, 7);
+    Assert.assertEquals(1, tabLine.getTab());
+    Assert.assertEquals(2, tabLine.getLineInTab());
+  }
+
+  @Test
+  public void getTabMiddleBackEdge() {
+    Assert.assertEquals(2, ProblemFactory.getTab(starts, 10).getTab());
+  }
+
+  @Test
+  public void getTabEnd() {
+    Assert.assertEquals(2, ProblemFactory.getTab(starts, 15).getTab());
   }
 
 }
